Adds utilities file.  Breaks repl out
