+++ index.js
@@ -1,9 +1,14 @@
+'use strict';
+
 const build = require('./build');
 const upload = require('./upload');
 const repl = require('./repl');
+const utils = require('./utils');
+
 
 module.exports = {
     build,
     upload,
+    utils,
     repl
 };
