+++ upload.js
@@ -1,39 +1,26 @@
 /* Upload file */
 'use strict';
 const espruino = require('espruino');
-const repl = require('./repl');
+const utils = require('./utils');
 
-/**
- * Espruino libraries use console.log directly.  This silences code in `fn` until the `callback` is called.
- */
-function mute(fn) {
-    let oldLog = console.log;
-    console.log = Function.prototype;
-    fn.call(null, () => {
-        console.log = oldLog;
-    });
-}
 
 function uploadToDevice(device, code) {
     console.log(`Sending code to device - ${device.port} @ ${device.baud_rate} baud...`);
-    mute((unmute) => {
+    utils.mute((unmute) => {
         espruino.init(() => {
             Espruino.Config.BAUD_RATE = device.baud_rate;
             espruino.sendCode(device.port, code, () => {
                 unmute();
                 console.log(`Code sent to ${device.port}`);
-                // This waits for the port to actually close, before listening
-                // FIXME:csd This is similar to what is happening in Espruino.  See "disconnected" processor.
-                setTimeout(() => repl(device), 2000);
+                // FIXME: Espruino is holding the process open
+                process.exit();
             });
         });
     });
 }
 
 module.exports = function upload(devices) {
-    let espruinoDevices = Object.keys(devices)
-                            .filter(port => devices[port].runtime === "espruino")
-                            .map(port => Object.assign({port}, devices[port]));
+    let espruinoDevices = utils.filterDevices(devices);
 
     return (code) => {
         espruinoDevices.forEach(device => uploadToDevice(device, code));
